+++ jaseci_core/jaseci/__init__.py
@@ -1,8 +1,8 @@
-import os
+from os.path import dirname, join
 
 
 def get_ver():
-    with open(os.path.join("../", "VERSION")) as version_file:
+    with open(join(dirname(__file__) + "/../../", "VERSION")) as version_file:
         return version_file.read().strip()